multi: Port upstream sat_per_vbyte work
This is a port of the upstream commit 9d0d88ac2. It reserves the fields
introduced upstream in the protobuf files as well as ports the code
changes in order to ease future porting effort.

+++ cmd/dcrlncli/walletrpc_active.go
@@ -608,7 +608,7 @@ func fundPsbt(ctx *cli.Context) error {
 
 	case ctx.Uint64("atoms_per_byte") > 0:
 		req.Fees = &walletrpc.FundPsbtRequest_AtomsPerByte{
-			AtomsPerByte: uint32(ctx.Uint64("atoms_per_byte")),
+			AtomsPerByte: ctx.Uint64("atoms_per_byte"),
 		}
 	}
